hypervisor: arm64: build the GICv2 code from the AArch32 port
We add to AArch64 the necessary stubs and missing variables that is needed
by the existing GICv2 implementation from AArch32.

GICv3 is still trickier; it makes heavy use of the sysregs, so we will
need to review more carefully how to port this code.

This is not really used by the hypervisor binary at the moment, but
building it is the first step nonetheless :)

Signed-off-by: Antonios Motakis <antonios.motakis@huawei.com>