hypervisor: arm64: add sysregs helper macros
On the AArch32 port of Jailhouse, the arm_write_sysreg and arm_read_sysreg
macros are used to access the system registers. Implementing these helpers
on AArch64 might allow for some low hanging fruit when looking for things
to reuse from AArch32.

On AArch64 we always get a u64 value from a system register, while on
AArch32 it might be either u32 or u64. We need to look at the users of
this macro on a case by case basis and see if it makes sense to reuse
code.

Signed-off-by: Antonios Motakis <antonios.motakis@huawei.com>