hypervisor: arm64: add definitions for the Aarch64 page table format
The Aarch64 page table format is very similar, almost identical to the
Aarch32 page table format. Add a header file for the Aarch64 page table
format, based on the Aarch32 implementation.

Aarch64 introduces an extra level of page tables, for a total of four, and
support for different translation granule sizes. Disabling the level zero
of the page tables, and using a granule size of 4Kb, results in an
identical page table format as Aarch32. With these parameters, we can
address an address range of 39 bits.

Signed-off-by: Antonios Motakis <antonios.motakis@huawei.com>