driver: ioremap the hypervisor firmware to any kernel address
At the moment the Linux driver maps the Jailhouse binary to
JAILHOUSE_BASE. The underlying assumption is that Linux may map the
firmware (in the Linux kernel space), to the same virtual address it has
been built to run from.

This assumption is unworkable on ARMv8 processors running in Aarch64 mode.
Kernel memory is allocated in a high address region, that is not
addressable from EL2, where the hypervisor will run from. This patch
removes the assumption, by introducing JAILHOUSE_IOMAP_ADDR. On current
targets, this is set to equal JAILHOUSE_BASE.

On Aarch64, JAILHOUSE_BASE will be the physical address the hypervisor
will be loaded to. This way, Jailhouse will run with identity mapping in
EL2. The Linux driver sets the address to the UART accordingly.

Signed-off-by: Antonios Motakis <antonios.motakis@huawei.com>