hypervisor: arm64: implement entry code to the hypervisor firmware
Implement the entry point we will jump to after the Linux driver
loads the firmware to memory. Here we also set up a stack for the
hypervisor to use.

Unlike AArch32, we jump to EL2 as soon as we enter the hypervisor
binary.

To do this, we also set up temporary MMU mappings. We use just two
pages, to statically configure the MMU for identity mapping;
we need this in order to perform unaligned accesses from the
hypervisor binary during early initialization.

To generate the early page tables at build time, we need to know the
Jailhouse physical address, and the physical address of the debug
UART. We introduce this change of behaviour in AArch64, because in
this architecture we do not have the option of using the page tables
set up previously by Linux.

Signed-off-by: Antonios Motakis <antonios.motakis@huawei.com>