hypervisor: arm64: add definitions for the AArch64 page table format
The AArch64 page table format is very similar, almost identical to
the AArch32 page table format. Add a header file for the AArch64 page
table format, based on the AArch32 implementation.

AArch64 introduces an extra level of page tables, for a total of
four, and support for different translation granule sizes. Disabling
the level zero of the page tables, and using a granule size of 4Kb,
results in an identical page table format with AArch32. With these
parameters, we can address 39 bits of memory.

Signed-off-by: Antonios Motakis <antonios.motakis@huawei.com>