arm64: initialize EL2 for hypervisor entry
Unlike AArch32, we jump to EL2 as soon as we enter the hypervisor
binary.

In order to do that, we also need to switch on the MMU, so we
generate early bootstrap page tables.
We need these in order to perform unaligned accesses from the
hypervisor binary during early initialization, and to avoid having
to maintain the caches during initialization.

Signed-off-by: Antonios Motakis <antonios.motakis@huawei.com>
Signed-off-by: Dmitry Voytik <dmitry.voytik@huawei.com>