arm64: add work around for large SPIs on AMD Seattle
The AMD Seattle board features SPI ids that are larger than 64,
which we do not support properly. This workaround allows us to
demonstrate working cells on this target, until we have a proper fix.

This implies that only specific setups will be used on the
AMD Seattle; the IRQs for the uart and the second xgmac are being
passed to the cells.

Signed-off-by: Antonios Motakis <antonios.motakis@huawei.com>